+++ src/main/java/pl/polsl/cargoflow/controller/PositionController.java
@@ -3,10 +3,10 @@ package pl.polsl.cargoflow.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import pl.polsl.cargoflow.model.dto.PositionRequest;
 import pl.polsl.cargoflow.model.dto.PositionResponse;
+import pl.polsl.cargoflow.model.exception.UnauthorizedException;
 import pl.polsl.cargoflow.service.AuthService;
 import pl.polsl.cargoflow.service.PositionService;
 import java.util.List;
@@ -26,35 +26,50 @@ public class PositionController {
 
     @GetMapping("/{id}")
     public ResponseEntity<PositionResponse> getPositionById(@RequestHeader("Authorization") String authHeader, @PathVariable Long id) {
-        authService.authenticate(authHeader);
+        boolean isAuthenticated = authService.authenticateAdmin(authHeader);
+        if (!isAuthenticated) {
+            throw new UnauthorizedException();
+        }
         PositionResponse positionResponse = positionService.getById(id);
         return ResponseEntity.ok(positionResponse);
     }
 
     @GetMapping
     public ResponseEntity<List<PositionResponse>> getAllPositions(@RequestHeader("Authorization") String authHeader) {
-        authService.authenticate(authHeader);
+        boolean isAuthenticated = authService.authenticateAdmin(authHeader);
+        if (!isAuthenticated) {
+            throw new UnauthorizedException();
+        }
         List<PositionResponse> positionResponses = positionService.getAll();
         return ResponseEntity.ok(positionResponses);
     }
 
     @PostMapping
     public ResponseEntity<PositionResponse> createPosition(@RequestHeader("Authorization") String authHeader, @RequestBody PositionRequest positionRequest) {
-        authService.authenticate(authHeader);
+        boolean isAuthenticated = authService.authenticateAdmin(authHeader);
+        if (!isAuthenticated) {
+            throw new UnauthorizedException();
+        }
         PositionResponse positionResponse = positionService.save(positionRequest);
         return ResponseEntity.status(HttpStatus.CREATED).body(positionResponse);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<PositionResponse> updatePosition(@RequestHeader("Authorization") String authHeader, @PathVariable Long id, @RequestBody PositionRequest positionRequest) {
-        authService.authenticate(authHeader);
+        boolean isAuthenticated = authService.authenticateAdmin(authHeader);
+        if (!isAuthenticated) {
+            throw new UnauthorizedException();
+        }
         PositionResponse positionResponse = positionService.update(id, positionRequest);
         return ResponseEntity.ok(positionResponse);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePosition(@RequestHeader("Authorization") String authHeader, @PathVariable Long id) {
-        authService.authenticate(authHeader);
+        boolean isAuthenticated = authService.authenticateAdmin(authHeader);
+        if (!isAuthenticated) {
+            throw new UnauthorizedException();
+        }
         positionService.delete(id);
         return ResponseEntity.noContent().build();
     }
